Reducers, actions and action constants moved to redux/ dir
