--- client/src/actions/hideModal.js
@@ -1,5 +1,0 @@
-import { HIDE_MODAL } from '../constants/actionConstants'
-
-export default _ => {
-    return { type: HIDE_MODAL }
-}
