+++ client/src/components/Header/components/UserIcon/UserIcon-Redux.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 
 import UserIconView from './UserIcon-View'
-import logoutUserAction from '../../../../actions/logoutUser'
+import { logoutUser as logoutUserAction } from '../../../../redux/actions/user-actions'
 
 const UserIconRedux = ({ username, logoutUser }) => {
     return <UserIconView username={username} logoutUser={logoutUser} />
