+++ client/src/components/hoc/WithModalTrigger/index.js
@@ -1,0 +1,3 @@
+import WithModalTrigger from './WithModalTrigger-Redux'
+
+export default WithModalTrigger
