+++ client/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { createStore, compose, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
-import reducers from './reducers'
+import reducers from './redux/reducers'
 import App from './components/App'
 import ModalContainer from './components/modals/ModalContainer'
