+++ client/src/redux/action-constants.js
@@ -1,5 +1,4 @@
 export const CONNECT_TO_SERVER = "CONNECT_TO_SERVER";
-export const RECIEVE_MESSAGE = "RECIEVE_MESSAGE";
 
 export const FETCH_PUBLIC_ROOMS = 'FETCH_PUBLIC_ROOMS'
 export const CREATE_ROOM = "CREATE_ROOM";
