+++ client/src/redux/actions/modal-actions/hideModal.js
@@ -1,0 +1,5 @@
+import { HIDE_MODAL } from '../../action-constants'
+
+export default _ => {
+    return { type: HIDE_MODAL }
+}
