+++ client/src/redux/actions/modal-actions/index.js
@@ -1,0 +1,4 @@
+import hideModal from './hideModal'
+import showModal from './showModal'
+
+export { hideModal, showModal }
