+++ client/src/redux/actions/notification-actions/createNotification.js
@@ -1,0 +1,9 @@
+import crypto from 'crypto'
+import { CREATE_NOTIFICATION } from '../../action-constants'
+
+export default (title, content, actions) => {
+    return {
+        type: CREATE_NOTIFICATION,
+        payload: { id: crypto.randomBytes(8), title, content, actions }
+    }
+}
