+++ client/src/redux/actions/notification-actions/index.js
@@ -1,0 +1,3 @@
+import createNotification from './createNotification'
+
+export { createNotification }
