+++ client/src/redux/actions/room-actions/fetchPublicRooms.js
@@ -1,5 +1,5 @@
-import { FETCH_PUBLIC_ROOMS } from '../constants/actionConstants'
-import { getAllRooms } from '../server-communication/httpServer'
+import { FETCH_PUBLIC_ROOMS } from '../../action-constants'
+import { getAllRooms } from '../../../server-communication/httpServer'
 
 export default (onFailure) => async dispatch => {
     getAllRooms((response) => {
