+++ client/src/redux/actions/room-actions/index.js
@@ -1,0 +1,7 @@
+import createRoom from './createRoom'
+import deleteRoom from './deleteRoom'
+import joinRoom from './joinRoom'
+import leaveRoom from './leaveRoom'
+import fetchPublicRooms from './fetchPublicRooms'
+
+export { createRoom, deleteRoom, joinRoom, leaveRoom, fetchPublicRooms }
