+++ client/src/redux/actions/room-list-sort-technique-actions/index.js
@@ -1,0 +1,3 @@
+import setRoomListSortTechnique from './setRoomListSortTechnique'
+
+export { setRoomListSortTechnique }
