+++ client/src/redux/actions/room-list-sort-technique-actions/setRoomListSortTechnique.js
@@ -1,4 +1,4 @@
-import { SET_ROOM_LIST_SORT_TECHNIQUE } from "../constants/actionConstants"
+import { SET_ROOM_LIST_SORT_TECHNIQUE } from "../../action-constants"
 
 export default (technique) => {
     return {
