+++ client/src/redux/actions/server-actions/index.js
@@ -1,0 +1,3 @@
+import connectToServer from './connectToServer'
+
+export { connectToServer }
