+++ client/src/redux/actions/user-actions/index.js
@@ -1,0 +1,6 @@
+import loginUser from './loginUser'
+import logoutUser from './logoutUser'
+import registerUser from './registerUser'
+import updateUser from './updateUser'
+
+export { loginUser, logoutUser, registerUser, updateUser }
