+++ client/src/redux/actions/user-actions/logoutUser.js
@@ -1,6 +1,6 @@
-import { LOGOUT } from '../constants/actionConstants';
+import { LOGOUT } from '../../action-constants';
 
-import { logoutUser } from '../server-communication/httpServer'
+import { logoutUser } from '../../../server-communication/httpServer'
 
 export default () => async dispatch => {
     logoutUser(() => dispatch({ type: LOGOUT }),({ response }) => console.log('Logout Failed: ', response));
