+++ client/src/redux/reducers/roomListSortTechniqueReducer.js
@@ -1,4 +1,4 @@
-import { SET_ROOM_LIST_SORT_TECHNIQUE } from "../constants/actionConstants"
+import { SET_ROOM_LIST_SORT_TECHNIQUE } from '../action-constants'
 
 const INITIAL_STATE = 'name-asc'
